Reorganize CIJoe::Server so that it can be used like a true Sinatra::Base.
In particular, it is now possible to mount Server at a subpath:

    map '/foo' do
       run CIJoe::Server.set(:project_path => 'projects/foo')
    end

and you can even run multiple instances of Joe if you subclass:

    map '/foo' do
      run Class.new(CIJoe::Server).set(:project_path => 'projects/foo')
    end
    map '/bar' do
      run Class.new(CIJoe::Server).set(:project_path => 'projects/bar')
    end

The only real change here is that we expose 'project_path' as a class variable,
which we can 'set'. We retain backwards compatibility (Server.start), so the
binary and other files all stay the same.

Signed-off-by: Chris Wanstrath <chris@ozmm.org>